+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/DataConnectSettings.kt
@@ -31,11 +31,21 @@ import java.util.Objects
  * @property sslEnabled Whether to use SSL for the connection; if `true`, then the connection will
  * be encrypted using SSL and, if false, the connection will _not_ be encrypted and all network
  * transmission will happen in plaintext.
+ * @property cacheSettings The local caching settings; if `null` then do not perform any local
+ * caching.
  */
 public class DataConnectSettings(
   public val host: String = "firebasedataconnect.googleapis.com",
-  public val sslEnabled: Boolean = true
+  public val sslEnabled: Boolean = true,
+  public val cacheSettings: CacheSettings?,
 ) {
+
+  // TODO(BreakingChange): Delete this constructor and set the default value for `cacheSettings`
+  //  in the primary constructor.
+  public constructor(
+    host: String = "firebasedataconnect.googleapis.com",
+    sslEnabled: Boolean = true,
+  ) : this(host = host, sslEnabled = sslEnabled, cacheSettings = null)
 
   /**
    * Compares this object with another object for equality.
@@ -46,7 +56,10 @@ public class DataConnectSettings(
    * object.
    */
   override fun equals(other: Any?): Boolean =
-    (other is DataConnectSettings) && other.host == host && other.sslEnabled == sslEnabled
+    (other is DataConnectSettings) &&
+      other.host == host &&
+      other.sslEnabled == sslEnabled &&
+      other.cacheSettings == cacheSettings
 
   /**
    * Calculates and returns the hash code for this object.
@@ -56,7 +69,8 @@ public class DataConnectSettings(
    * @return the hash code for this object, that incorporates the values of this object's public
    * properties.
    */
-  override fun hashCode(): Int = Objects.hash(DataConnectSettings::class, host, sslEnabled)
+  override fun hashCode(): Int =
+    Objects.hash(DataConnectSettings::class, host, sslEnabled, cacheSettings)
 
   /**
    * Returns a string representation of this object, useful for debugging.
@@ -70,13 +84,25 @@ public class DataConnectSettings(
    * @return a string representation of this object, which includes the class name and the values of
    * all public properties.
    */
-  override fun toString(): String = "DataConnectSettings(host=$host, sslEnabled=$sslEnabled)"
+  override fun toString(): String =
+    "DataConnectSettings(host=$host, sslEnabled=$sslEnabled, cacheSettings=$cacheSettings)"
 }
+
+/** Creates and returns a new [DataConnectSettings] instance with the given property values. */
+// TODO(BreakingChange): Delete this method and set the default value for `cacheSettings` in the
+//  remaining copy() method.
+public fun DataConnectSettings.copy(
+  host: String = this.host,
+  sslEnabled: Boolean = this.sslEnabled,
+): DataConnectSettings =
+  DataConnectSettings(host = host, sslEnabled = sslEnabled, cacheSettings = cacheSettings)
 
 /** Creates and returns a new [DataConnectSettings] instance with the given property values. */
 public fun DataConnectSettings.copy(
   host: String = this.host,
-  sslEnabled: Boolean = this.sslEnabled
-): DataConnectSettings = DataConnectSettings(host = host, sslEnabled = sslEnabled)
+  sslEnabled: Boolean = this.sslEnabled,
+  cacheSettings: CacheSettings?,
+): DataConnectSettings =
+  DataConnectSettings(host = host, sslEnabled = sslEnabled, cacheSettings = cacheSettings)
 
 internal fun DataConnectSettings.isDefaultHost() = host == DataConnectSettings().host